requirements.txt
asgiref==3.3.4
Django==3.2.3
django-crispy-forms==1.11.2
django-debug-toolbar==3.2.1
django-shopping-cart==0.1
Pillow==8.2.0
psycopg2-binary==2.8.6
pytz==2021.1
sqlparse==0.4.1
